Move rest of entities to correct packages/modules
