+++ ippon/fight/views.py
@@ -1,5 +1,5 @@
 from rest_framework import viewsets, permissions
-from rest_framework.decorators import action
+from rest_framework.decorators import action, api_view
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
@@ -9,6 +9,7 @@ import ippon.models
 import ippon.models.fight
 import ippon.point.serializers as pts
 import ippon.models.point as ptm
+from ippon.tournament.authorizations import has_tournament_authorization
 
 
 class FightViewSet(viewsets.ModelViewSet):
@@ -25,3 +26,9 @@ class FightViewSet(viewsets.ModelViewSet):
         fight = get_object_or_404(self.queryset, pk=pk)
         serializer = pts.PointSerializer(ptm.Point.objects.filter(fight=fight), many=True)
         return Response(serializer.data)
+
+
+@api_view(['GET'])
+def fight_authorization(request, pk, format=None):
+    fight = ippon.models.fight.Fight.objects.get(pk=pk)
+    return has_tournament_authorization([True, False], fight.team_fight.tournament.id, request)
