+++ ippon/urls.py
@@ -3,7 +3,6 @@ from django.urls import include
 from rest_framework.routers import DefaultRouter
 from rest_framework.schemas import get_schema_view
 
-import ippon.authorization_views
 import ippon.club.views
 import ippon.cup_fight.views
 import ippon.cup_phase.views
@@ -43,25 +42,25 @@ urlpatterns = [
     url(r'^', include(router.urls)),
     url(r'user-data/', ippon.user.views.user_data, name="user-data"),
     url(r'^schema/', schema_view),
-    url(r'^authorization/clubs/(?P<pk>[0-9]+)/$', ippon.authorization_views.club_authorization,
+    url(r'^authorization/clubs/(?P<pk>[0-9]+)/$', ippon.club.views.club_authorization,
         name='club-authorization'),
-    url(r'^authorization/tournaments/staff/(?P<pk>[0-9]+)/$', ippon.authorization_views.tournament_staff_authorization,
+    url(r'^authorization/tournaments/staff/(?P<pk>[0-9]+)/$', ippon.tournament.views.tournament_staff_authorization,
         name='tournament-staff-authorization'),
-    url(r'^authorization/tournaments/admins/(?P<pk>[0-9]+)/$', ippon.authorization_views.tournament_admin_authorization,
+    url(r'^authorization/tournaments/admins/(?P<pk>[0-9]+)/$', ippon.tournament.views.tournament_admin_authorization,
         name='tournament-admin-authorization'),
-    url(r'^authorization/fights/(?P<pk>[0-9]+)/$', ippon.authorization_views.fight_authorization,
+    url(r'^authorization/fights/(?P<pk>[0-9]+)/$', ippon.fight.views.fight_authorization,
         name='fight-authorization'),
-    url(r'^authorization/team_fights/(?P<pk>[0-9]+)/$', ippon.authorization_views.team_fight_authorization,
+    url(r'^authorization/team_fights/(?P<pk>[0-9]+)/$', ippon.team_fight.views.team_fight_authorization,
         name='team-fight-authorization'),
-    url(r'^authorization/teams/(?P<pk>[0-9]+)/$', ippon.authorization_views.team_authorization,
+    url(r'^authorization/teams/(?P<pk>[0-9]+)/$', ippon.team.views.team_authorization,
         name='team-authorization'),
-    url(r'^authorization/groups/(?P<pk>[0-9]+)/$', ippon.authorization_views.group_authorization,
+    url(r'^authorization/groups/(?P<pk>[0-9]+)/$', ippon.group.views.group_authorization,
         name='group-authorization'),
     url(r'^authorization/group_phases/(?P<pk>[0-9]+)/$', ippon.group_phase.views.group_phase_authorization,
         name='group-phase-authorization'),
     url(r'^authorization/cup_phases/(?P<pk>[0-9]+)/$', ippon.cup_phase.views.cup_phase_authorization,
         name='cup-phase-authorization'),
-    url(r'^authorization/players/(?P<pk>[0-9]+)/$', ippon.authorization_views.player_authorization,
+    url(r'^authorization/players/(?P<pk>[0-9]+)/$', ippon.player.views.player_authorization,
         name='player-authorization'),
     url(r'^registration/', ippon.user.views.register_user, name='register-user'),
     url(r'^shallow_players/(?P<pk>[0-9]+)/$', ippon.player.views.ShallowPlayerDetailView.as_view(),
